Cleanup AST/CFG Dot Generators
